akash-provider: use dynamic fees
Previously AKASH_FEES was set to 5123uakt which was working for the
provider but sporadically running "out of gas" for "akash tx provider
create" / "akash tx cert create server" transactions.

As the first attempt to address this, the fees were jacked up to 40000uakt.

However, since AKASH_FEES was slipping through to akash provider due
to it being a global environment variable, this was causing provider to use
way too high fees for all txs.

The core problem was that the --gas and --gas-adjustment parameters were
ignored as we do not pre-seed 'minimum-gas-prices = "0.025uakt"' via
/.akash/config/app.toml file like it suggested in the Akash docs.

To address this, I propsoe introducing AKASH_MINIMUM_GAS_PRICES "back again"
in order to use dynamic fees.

This will also likey to help in resolving this issue
https://github.com/ovrclk/akash/issues/1508