Create price_script_whitelist.sh
Non-audited price script that makes use of `$AKASH_OWNER`.

The script makes it possible for providers to only accept tenants that they trust. This can be tenant addresses from their own company, simply using Akash privately, or if they wish to only rent out to users they have, for example, done KYC on.

A file in the same directory, `whitelist.txt`, is required. Line by line you add your whitelisted addresses with the bech32 format to add to the whitelist.

Everything works when I test the script locally, but I am not a provider so I would appreciate a code review and someone to test it.

https://github.com/ovrclk/akash/issues/1617